refactor(config): isolate runtime backend policy interpretation
