+++ crates/config/src/runtime/policies/backend.rs
@@ -1,1 +1,173 @@
-// Skeleton module for runtime backend policy interpretation.
+use std::time::Duration;
+
+use super::config_invalid;
+use crate::{
+    backend_endpoint::{BackendEndpoint, BackendScheme},
+    config::{HealthCheck, Performance, UpstreamTls},
+    runtime::RuntimeConfigError,
+};
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum RuntimeBackendAddressKind {
+    Hostname,
+    IpLiteral,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct RuntimeBackendEndpoint {
+    pub configured_address: String,
+    pub canonical: BackendEndpoint,
+    pub origin: String,
+    pub authority_host: String,
+    pub authority_port: u16,
+    pub address_kind: RuntimeBackendAddressKind,
+    pub transport_kind: super::RuntimeBackendTransportKind,
+}
+
+impl RuntimeBackendEndpoint {
+    pub(crate) fn normalize(
+        upstream_name: &str,
+        backend_id: &str,
+        address: &str,
+    ) -> Result<Self, RuntimeConfigError> {
+        let canonical = BackendEndpoint::parse(address).map_err(|reason| {
+            RuntimeConfigError::BackendAddressInvalid {
+                upstream: upstream_name.to_string(),
+                backend: backend_id.to_string(),
+                address: address.to_string(),
+                reason,
+            }
+        })?;
+        let authority_host = canonical.authority_host().to_string();
+        let authority_port = canonical.authority_port();
+        let address_kind = if canonical.authority_is_ip_literal() {
+            RuntimeBackendAddressKind::IpLiteral
+        } else {
+            RuntimeBackendAddressKind::Hostname
+        };
+        let transport_kind = match canonical.scheme() {
+            BackendScheme::Http => super::RuntimeBackendTransportKind::Http1,
+            BackendScheme::Https => super::RuntimeBackendTransportKind::H2,
+        };
+        let origin = canonical.origin();
+
+        Ok(Self {
+            configured_address: address.to_string(),
+            canonical,
+            origin,
+            authority_host,
+            authority_port,
+            address_kind,
+            transport_kind,
+        })
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct RuntimeBackendHealthCheck {
+    pub path: String,
+    pub interval: Duration,
+    pub timeout: Duration,
+    pub failure_threshold: u32,
+    pub success_threshold: u32,
+    pub cooldown: Duration,
+}
+
+impl RuntimeBackendHealthCheck {
+    pub(crate) fn normalize(
+        upstream_name: &str,
+        backend_id: &str,
+        health_check: &HealthCheck,
+    ) -> Result<Self, RuntimeConfigError> {
+        if health_check.interval == 0 {
+            return Err(config_invalid(format!(
+                "health check interval is invalid (0) for backend '{backend_id}' in upstream '{upstream_name}'"
+            )));
+        }
+        if health_check.timeout_ms == 0 {
+            return Err(config_invalid(format!(
+                "health check timeout is invalid (0) for backend '{backend_id}' in upstream '{upstream_name}'"
+            )));
+        }
+        if health_check.failure_threshold == 0 {
+            return Err(config_invalid(format!(
+                "health check failure threshold is invalid (0) for backend '{backend_id}' in upstream '{upstream_name}'"
+            )));
+        }
+        if health_check.success_threshold == 0 {
+            return Err(config_invalid(format!(
+                "health check success threshold is invalid (0) for backend '{backend_id}' in upstream '{upstream_name}'"
+            )));
+        }
+
+        Ok(Self {
+            path: if health_check.path.trim().is_empty() {
+                "/".to_string()
+            } else {
+                health_check.path.clone()
+            },
+            interval: Duration::from_millis(health_check.interval),
+            timeout: Duration::from_millis(health_check.timeout_ms),
+            failure_threshold: health_check.failure_threshold,
+            success_threshold: health_check.success_threshold,
+            cooldown: Duration::from_millis(health_check.cooldown_ms),
+        })
+    }
+
+    #[cfg(test)]
+    pub(crate) fn as_config(&self) -> HealthCheck {
+        HealthCheck {
+            path: self.path.clone(),
+            interval: self.interval.as_millis().try_into().unwrap_or(u64::MAX),
+            timeout_ms: self.timeout.as_millis().try_into().unwrap_or(u64::MAX),
+            failure_threshold: self.failure_threshold,
+            success_threshold: self.success_threshold,
+            cooldown_ms: self.cooldown.as_millis().try_into().unwrap_or(u64::MAX),
+        }
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct RuntimeBackendTlsPolicy {
+    pub verify_certificates: bool,
+    pub strict_sni: bool,
+    pub ca_file: Option<String>,
+    pub ca_dir: Option<String>,
+}
+
+impl RuntimeBackendTlsPolicy {
+    pub fn from_effective_tls(effective_tls: &UpstreamTls) -> Self {
+        Self {
+            verify_certificates: effective_tls.verify_certificates,
+            strict_sni: effective_tls.strict_sni,
+            ca_file: effective_tls.ca_file.clone(),
+            ca_dir: effective_tls.ca_dir.clone(),
+        }
+    }
+
+    pub fn as_upstream_tls(&self) -> UpstreamTls {
+        UpstreamTls {
+            verify_certificates: self.verify_certificates,
+            strict_sni: self.strict_sni,
+            ca_file: self.ca_file.clone(),
+            ca_dir: self.ca_dir.clone(),
+        }
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct RuntimeBackendDnsPolicy {
+    pub refresh_enabled: bool,
+    pub refresh_interval: Duration,
+}
+
+impl RuntimeBackendDnsPolicy {
+    pub(crate) fn from_performance(performance: &Performance) -> Self {
+        Self {
+            refresh_enabled: performance.backend_dns_refresh_enabled,
+            refresh_interval: Duration::from_millis(
+                performance.backend_dns_refresh_interval_ms,
+            ),
+        }
+    }
+}
